mix/backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Async Tasks
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Ansible Integration
ansible==9.2.0
ansible-runner==2.3.4

# Security
django-ratelimit==4.1.0
python-decouple==3.8

# Monitoring & Logging
django-prometheus==2.3.1
sentry-sdk==1.40.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0

# Development
django-debug-toolbar==4.3.0
django-extensions==3.2.3

# Server
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
